+++ src/Flutter.Material/FloatingActionButton.cs
@@ -2,6 +2,7 @@ using Avalonia;
 using Avalonia.Media;
 using Flutter.Foundation;
 using Flutter.Rendering;
+using Flutter.UI;
 using Flutter.Widgets;
 
 namespace Flutter.Material;
@@ -22,6 +23,8 @@ public sealed class FloatingActionButton : StatelessWidget
         Widget? child,
         Action? onPressed,
         string? tooltip = null,
+        object? heroTag = null,
+        MouseCursor? mouseCursor = null,
         Color? foregroundColor = null,
         Color? backgroundColor = null,
         Color? focusColor = null,
@@ -35,6 +38,8 @@ public sealed class FloatingActionButton : StatelessWidget
         bool mini = false,
         BorderRadius? shape = null,
         FocusNode? focusNode = null,
+        bool? enableFeedback = null,
+        Clip clipBehavior = Clip.None,
         bool autofocus = false,
         MaterialTapTargetSize? materialTapTargetSize = null,
         Key? key = null) : this(
@@ -44,6 +49,8 @@ public sealed class FloatingActionButton : StatelessWidget
             onPressed: onPressed,
             type: mini ? FloatingActionButtonType.Small : FloatingActionButtonType.Regular,
             isExtended: false,
+            heroTag: heroTag,
+            mouseCursor: mouseCursor,
             foregroundColor: foregroundColor,
             backgroundColor: backgroundColor,
             focusColor: focusColor,
@@ -56,6 +63,8 @@ public sealed class FloatingActionButton : StatelessWidget
             disabledElevation: disabledElevation,
             shape: shape,
             focusNode: focusNode,
+            enableFeedback: enableFeedback,
+            clipBehavior: clipBehavior,
             autofocus: autofocus,
             materialTapTargetSize: materialTapTargetSize,
             extendedIconLabelSpacing: null,
@@ -72,6 +81,8 @@ public sealed class FloatingActionButton : StatelessWidget
         Action? onPressed,
         FloatingActionButtonType type,
         bool isExtended,
+        object? heroTag,
+        MouseCursor? mouseCursor,
         Color? foregroundColor,
         Color? backgroundColor,
         Color? focusColor,
@@ -84,6 +95,8 @@ public sealed class FloatingActionButton : StatelessWidget
         double? disabledElevation,
         BorderRadius? shape,
         FocusNode? focusNode,
+        bool? enableFeedback,
+        Clip clipBehavior,
         bool autofocus,
         MaterialTapTargetSize? materialTapTargetSize,
         double? extendedIconLabelSpacing,
@@ -112,6 +125,8 @@ public sealed class FloatingActionButton : StatelessWidget
         OnPressed = onPressed;
         Type = type;
         IsExtended = isExtended;
+        HeroTag = heroTag;
+        MouseCursor = mouseCursor;
         ForegroundColor = foregroundColor;
         BackgroundColor = backgroundColor;
         FocusColor = focusColor;
@@ -124,6 +139,8 @@ public sealed class FloatingActionButton : StatelessWidget
         DisabledElevation = disabledElevation;
         Shape = shape;
         FocusNode = focusNode;
+        EnableFeedback = enableFeedback;
+        ClipBehavior = clipBehavior;
         Autofocus = autofocus;
         MaterialTapTargetSize = materialTapTargetSize;
         ExtendedIconLabelSpacing = extendedIconLabelSpacing;
@@ -142,6 +159,10 @@ public sealed class FloatingActionButton : StatelessWidget
     private FloatingActionButtonType Type { get; }
 
     public bool IsExtended { get; }
+
+    public object? HeroTag { get; }
+
+    public MouseCursor? MouseCursor { get; }
 
     public Color? ForegroundColor { get; }
 
@@ -167,6 +188,10 @@ public sealed class FloatingActionButton : StatelessWidget
 
     public FocusNode? FocusNode { get; }
 
+    public bool? EnableFeedback { get; }
+
+    public Clip ClipBehavior { get; }
+
     public bool Autofocus { get; }
 
     public MaterialTapTargetSize? MaterialTapTargetSize { get; }
@@ -181,6 +206,8 @@ public sealed class FloatingActionButton : StatelessWidget
         Widget? child,
         Action? onPressed,
         string? tooltip = null,
+        object? heroTag = null,
+        MouseCursor? mouseCursor = null,
         Color? foregroundColor = null,
         Color? backgroundColor = null,
         Color? focusColor = null,
@@ -193,6 +220,8 @@ public sealed class FloatingActionButton : StatelessWidget
         double? disabledElevation = null,
         BorderRadius? shape = null,
         FocusNode? focusNode = null,
+        bool? enableFeedback = null,
+        Clip clipBehavior = Clip.None,
         bool autofocus = false,
         MaterialTapTargetSize? materialTapTargetSize = null,
         Key? key = null)
@@ -204,6 +233,8 @@ public sealed class FloatingActionButton : StatelessWidget
             onPressed: onPressed,
             type: FloatingActionButtonType.Small,
             isExtended: false,
+            heroTag: heroTag,
+            mouseCursor: mouseCursor,
             foregroundColor: foregroundColor,
             backgroundColor: backgroundColor,
             focusColor: focusColor,
@@ -216,6 +247,8 @@ public sealed class FloatingActionButton : StatelessWidget
             disabledElevation: disabledElevation,
             shape: shape,
             focusNode: focusNode,
+            enableFeedback: enableFeedback,
+            clipBehavior: clipBehavior,
             autofocus: autofocus,
             materialTapTargetSize: materialTapTargetSize,
             extendedIconLabelSpacing: null,
@@ -228,6 +261,8 @@ public sealed class FloatingActionButton : StatelessWidget
         Widget? child,
         Action? onPressed,
         string? tooltip = null,
+        object? heroTag = null,
+        MouseCursor? mouseCursor = null,
         Color? foregroundColor = null,
         Color? backgroundColor = null,
         Color? focusColor = null,
@@ -240,6 +275,8 @@ public sealed class FloatingActionButton : StatelessWidget
         double? disabledElevation = null,
         BorderRadius? shape = null,
         FocusNode? focusNode = null,
+        bool? enableFeedback = null,
+        Clip clipBehavior = Clip.None,
         bool autofocus = false,
         MaterialTapTargetSize? materialTapTargetSize = null,
         Key? key = null)
@@ -251,6 +288,8 @@ public sealed class FloatingActionButton : StatelessWidget
             onPressed: onPressed,
             type: FloatingActionButtonType.Large,
             isExtended: false,
+            heroTag: heroTag,
+            mouseCursor: mouseCursor,
             foregroundColor: foregroundColor,
             backgroundColor: backgroundColor,
             focusColor: focusColor,
@@ -263,6 +302,8 @@ public sealed class FloatingActionButton : StatelessWidget
             disabledElevation: disabledElevation,
             shape: shape,
             focusNode: focusNode,
+            enableFeedback: enableFeedback,
+            clipBehavior: clipBehavior,
             autofocus: autofocus,
             materialTapTargetSize: materialTapTargetSize,
             extendedIconLabelSpacing: null,
@@ -277,6 +318,8 @@ public sealed class FloatingActionButton : StatelessWidget
         Widget? icon = null,
         bool isExtended = true,
         string? tooltip = null,
+        object? heroTag = null,
+        MouseCursor? mouseCursor = null,
         Color? foregroundColor = null,
         Color? backgroundColor = null,
         Color? focusColor = null,
@@ -289,6 +332,8 @@ public sealed class FloatingActionButton : StatelessWidget
         double? disabledElevation = null,
         BorderRadius? shape = null,
         FocusNode? focusNode = null,
+        bool? enableFeedback = null,
+        Clip clipBehavior = Clip.None,
         bool autofocus = false,
         MaterialTapTargetSize? materialTapTargetSize = null,
         double? extendedIconLabelSpacing = null,
@@ -303,6 +348,8 @@ public sealed class FloatingActionButton : StatelessWidget
             onPressed: onPressed,
             type: FloatingActionButtonType.Extended,
             isExtended: isExtended,
+            heroTag: heroTag,
+            mouseCursor: mouseCursor,
             foregroundColor: foregroundColor,
             backgroundColor: backgroundColor,
             focusColor: focusColor,
@@ -315,6 +362,8 @@ public sealed class FloatingActionButton : StatelessWidget
             disabledElevation: disabledElevation,
             shape: shape,
             focusNode: focusNode,
+            enableFeedback: enableFeedback,
+            clipBehavior: clipBehavior,
             autofocus: autofocus,
             materialTapTargetSize: materialTapTargetSize,
             extendedIconLabelSpacing: extendedIconLabelSpacing,
@@ -405,6 +454,8 @@ public sealed class FloatingActionButton : StatelessWidget
             onPressed: OnPressed,
             style: style,
             focusNode: FocusNode,
+            clipBehavior: ClipBehavior,
+            enableFeedback: EnableFeedback,
             autofocus: Autofocus);
 
         if (!string.IsNullOrWhiteSpace(Tooltip))